Lock to a specific version
It's probably better to minimize the impact of linter changes on the
project as a whole.

Lock the linter to a specific version so we don't accidentally hit a
linter upgrade which breaks all of our existing builds.

We can then run a separate fixup branch which upgrades the linter and
then fixes all the newly found errors.