Localize use of waitgroups
The code passed waitgroup/sizedwaitgroup to another routine which then
called wg.Done(). This breaks locality of waitgroups over two files,
and two functions.

Rewrite the code by using an anonymous function to call wg.Done(). The
result is that waitgroups are now local to the caller, and the callee
is out of the loop.